Added second order diffraction option to Monochromator_JY
Added second order diffraction option to Monochromator_JY. This could be expanded to higher orders by adding larger numbers to the diffractionOrder popup menu (pum). There was some minor rearranging of GUI elements to fit the pum. The properties "absoluteWavelenth" and "absoluteWavenumbers" were add to allow for the access of the absolute setting of the monochromator without considering the diffraction order. Now, the "wavelength" and "wavenumbers" properties will report the values of the diffraction order corrected values, e.g. when the monochromator is set for second order diffraction at 2350 cm-1 the absoluteWavenumbers property will read 4700 cm-1 and the wavenumbers property will read 2350 cm-1. If the diffraction order is set the 1, the wavenumbers and absoluteWavenumbers properties should be equal as the diffraction order correction is 1.